Renamed all files with pattern "_functions".
