+++ include/common/lock.smode.h
