+++ include/common/tablewalk.smode.h
