+++ src/common/heap.smode.c
@@ -2,10 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "heap_functions.smode.h"
+#include "heap.smode.h"
+#include "jumpstart.h"
 #include "jumpstart_defines.h"
-#include "jumpstart_functions.h"
-#include "lock_functions.smode.h"
+#include "lock.smode.h"
 
 #include <stdint.h>
