+++ src/common/jumpstart.smode.S
