+++ src/common/string.smode.c
