+++ src/public/uart/meson.build
@@ -3,6 +3,6 @@
 # SPDX-License-Identifier: Apache-2.0
 
 jumpstart_public_drivers += files(
-                          'uart_functions.mmode.c',
-                          'uart_functions.smode.c',
+                          'uart.mmode.c',
+                          'uart.smode.c',
                           )
