+++ tests/common/test002.c
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "cpu_bits.h"
-#include "jumpstart_functions.h"
+#include "jumpstart.h"
 
 // user mode functions
 // The assembly functions are already tagged with the .text.umode section
