+++ tests/common/test003.c
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "cpu_bits.h"
-#include "jumpstart_functions.h"
+#include "jumpstart.h"
 
 void test003_illegal_instruction_handler(void);
 int test003_illegal_instruction_function(void);
