+++ tests/common/test017.c
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "cpu_bits.h"
-#include "jumpstart_functions.h"
+#include "jumpstart.h"
 
 void test017_illegal_instruction_handler(void);
 int test017_illegal_instruction_function(void);
