+++ tests/common/test019.c
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "jumpstart_functions.h"
+#include "jumpstart.h"
 
 int main(void) {
   for (int i = 0; i < 10; ++i) {
