+++ tests/common/test020.c
@@ -3,8 +3,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "cpu_bits.h"
-#include "jumpstart_functions.h"
-#include "tablewalk_functions.smode.h"
+#include "jumpstart.h"
+#include "tablewalk.smode.h"
 
 int main(void) {
   if (get_field(read_csr(satp), SATP64_MODE) != VM_1_10_SV39) {
