+++ tests/common/test023.c
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "cpu_bits.h"
-#include "jumpstart_functions.h"
+#include "jumpstart.h"
 
 // Supervisor functions
 // The assembly functions are already tagged with the .text.smode section
